+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   [#111](https://github.com/TNO-S3/WuppieFuzz/pull/111)
 - Generate example parameters for the OpenAPI `allOf` keyword with just a
   single schema in [#118](https://github.com/TNO-S3/WuppieFuzz/pull/118)
+- Support logging in the output-corpus subcommand in [#119](https://github.com/TNO-S3/WuppieFuzz/pull/119)
 
 ## Fixes